refactor(tool): extract disallowed AST YAML rule model to own library
Move DisallowedAstMatchKind, DisallowedPatternRule, DisallowedAstViolation, and
parseDisallowedAstRulesFromYaml into tool/disallowed_ast_pattern_rules.dart so
check_disallowed_ast_patterns.dart holds scan/visitor only. Re-export rule types
from the checker for existing tests. Document in SPEC and repo-lint source table
(GitHub #1912 refactor slice).
