+++ tool/check_disallowed_ast_patterns.dart
@@ -9,6 +9,9 @@ import 'package:path/path.dart' as p;
 import 'package:yaml/yaml.dart';
 
 import 'ct_repo_lint_scan_contract.dart';
+import 'disallowed_ast_pattern_rules.dart';
+
+export 'disallowed_ast_pattern_rules.dart';
 
 /// PR-blocking structural AST checks driven by [tool/disallowed_ast_patterns.yaml].
 ///
@@ -32,7 +35,9 @@ int runCheckDisallowedAstPatterns(
     return 1;
   }
 
-  final allRules = _loadRules(configFile.readAsStringSync());
+  final allRules = parseDisallowedAstRulesFromYaml(
+    loadYaml(configFile.readAsStringSync()),
+  );
   final rules = enabledRuleIds == null
       ? allRules
       : allRules.where((rule) => enabledRuleIds.contains(rule.id)).toList();
@@ -164,329 +169,7 @@ List<DisallowedAstViolation> findDisallowedAstViolations(
 }
 
 List<DisallowedPatternRule> loadDisallowedAstRulesForTest(String yamlText) =>
-    _parseRulesYaml(loadYaml(yamlText));
-
-List<DisallowedPatternRule> _loadRules(String yamlText) =>
-    _parseRulesYaml(loadYaml(yamlText));
-
-List<DisallowedPatternRule> _parseRulesYaml(Object? yamlRoot) {
-  if (yamlRoot is! YamlMap) {
-    return const [];
-  }
-  final rulesNode = yamlRoot['rules'];
-  if (rulesNode is! YamlList) {
-    return const [];
-  }
-  final out = <DisallowedPatternRule>[];
-  for (final entry in rulesNode.nodes) {
-    final value = entry.value;
-    if (value is! YamlMap) {
-      continue;
-    }
-    final id = value['id']?.toString();
-    final message = value['message']?.toString().trim();
-    final match = value['match'];
-    if (id == null || id.isEmpty || message == null || message.isEmpty) {
-      continue;
-    }
-    if (match is! YamlMap) {
-      continue;
-    }
-    final kind = match['kind']?.toString();
-    if (kind == 'cascaded_method_invocation') {
-      final namesNode = match['method_names'];
-      if (namesNode is! YamlList) {
-        continue;
-      }
-      final names = <String>{};
-      for (final n in namesNode.nodes) {
-        final s = n.value?.toString();
-        if (s != null && s.isNotEmpty) {
-          names.add(s);
-        }
-      }
-      if (names.isEmpty) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.cascadedMethodInvocation,
-          cascadedMethodNames: names,
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'stream_where_is_map_as') {
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.streamWhereIsMapAs,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'comment_substring') {
-      final needle = match['contains']?.toString();
-      if (needle == null || needle.isEmpty) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.commentSubstring,
-          cascadedMethodNames: const {},
-          commentSubstring: needle,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'raw_named_type') {
-      final namesNode = match['type_names'];
-      if (namesNode is! YamlList) {
-        continue;
-      }
-      final names = <String>{};
-      for (final n in namesNode.nodes) {
-        final s = n.value?.toString();
-        if (s != null && s.isNotEmpty) {
-          names.add(s);
-        }
-      }
-      if (names.isEmpty) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.rawNamedType,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: names,
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'method_body_line_span') {
-      final functionName = match['function_name']?.toString();
-      final maxBodyLineSpan = int.tryParse(
-        match['max_body_line_span']?.toString() ?? '',
-      );
-      final requireWidgetClassExtends =
-          match['require_widget_class_extends'] == true;
-      if (functionName == null ||
-          functionName.isEmpty ||
-          maxBodyLineSpan == null ||
-          maxBodyLineSpan < 1) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.methodBodyLineSpan,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: functionName,
-          maxBodyLineSpan: maxBodyLineSpan,
-          requireWidgetClassExtends: requireWidgetClassExtends,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'unprefixed_province_id_string_literal_argument') {
-      final namesNode = match['method_names'];
-      final argumentIndex = int.tryParse(
-        match['argument_index']?.toString() ?? '',
-      );
-      if (namesNode is! YamlList ||
-          argumentIndex == null ||
-          argumentIndex < 0) {
-        continue;
-      }
-      final names = <String>{};
-      for (final n in namesNode.nodes) {
-        final s = n.value?.toString();
-        if (s != null && s.isNotEmpty) {
-          names.add(s);
-        }
-      }
-      if (names.isEmpty) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind:
-              DisallowedAstMatchKind.unprefixedProvinceIdStringLiteralArgument,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: argumentIndex,
-          invocationMethodNames: names,
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'sea_zone_local_id_extraction') {
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.seaZoneLocalIdExtraction,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'sea_zone_bucket_lookup_without_canonical_key') {
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.seaZoneBucketLookupWithoutCanonicalKey,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'province_local_segment_boundary_only') {
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.provinceLocalSegmentBoundaryOnly,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: const {},
-          packageName: null,
-          allowedPackageImports: const {},
-        ),
-      );
-    } else if (kind == 'scoped_package_import_contract') {
-      final scopeNode = match['scoped_relative_path_prefixes'];
-      final packageName = match['package_name']?.toString();
-      final allowedImportsYamlList = match['allowed_imports'];
-      if (scopeNode is! YamlList ||
-          packageName == null ||
-          packageName.isEmpty ||
-          allowedImportsYamlList is! YamlList) {
-        continue;
-      }
-      final scopedRelativePathPrefixes = <String>{};
-      for (final scopeEntry in scopeNode.nodes) {
-        final scope = scopeEntry.value?.toString();
-        if (scope != null && scope.isNotEmpty) {
-          scopedRelativePathPrefixes.add(scope);
-        }
-      }
-      final allowedPackageImports = <String>{};
-      for (final importUriEntry in allowedImportsYamlList.nodes) {
-        final importPath = importUriEntry.value?.toString();
-        if (importPath != null && importPath.isNotEmpty) {
-          allowedPackageImports.add(importPath);
-        }
-      }
-      if (scopedRelativePathPrefixes.isEmpty || allowedPackageImports.isEmpty) {
-        continue;
-      }
-      out.add(
-        DisallowedPatternRule(
-          id: id,
-          message: message,
-          kind: DisallowedAstMatchKind.scopedPackageImportContract,
-          cascadedMethodNames: const {},
-          commentSubstring: null,
-          rawNamedTypeNames: const {},
-          functionName: null,
-          maxBodyLineSpan: null,
-          requireWidgetClassExtends: false,
-          argumentIndex: null,
-          invocationMethodNames: const {},
-          allowedRelativePaths: const {},
-          scopedRelativePathPrefixes: scopedRelativePathPrefixes,
-          packageName: packageName,
-          allowedPackageImports: allowedPackageImports,
-        ),
-      );
-    }
-  }
-  return out;
-}
+    parseDisallowedAstRulesFromYaml(loadYaml(yamlText));
 
 bool _fileIgnoresRule(
   String source,
@@ -522,70 +205,6 @@ bool _isSuppressedAtLine(
     return true;
   }
   return false;
-}
-
-/// Kinds of structural matches defined in [tool/disallowed_ast_patterns.yaml].
-enum DisallowedAstMatchKind {
-  cascadedMethodInvocation,
-  streamWhereIsMapAs,
-  commentSubstring,
-  rawNamedType,
-  methodBodyLineSpan,
-  seaZoneLocalIdExtraction,
-  seaZoneBucketLookupWithoutCanonicalKey,
-  unprefixedProvinceIdStringLiteralArgument,
-  provinceLocalSegmentBoundaryOnly,
-  scopedPackageImportContract,
-}
-
-class DisallowedPatternRule {
-  const DisallowedPatternRule({
-    required this.id,
-    required this.message,
-    required this.kind,
-    required this.cascadedMethodNames,
-    required this.commentSubstring,
-    required this.rawNamedTypeNames,
-    required this.functionName,
-    required this.maxBodyLineSpan,
-    required this.requireWidgetClassExtends,
-    required this.argumentIndex,
-    required this.invocationMethodNames,
-    required this.allowedRelativePaths,
-    required this.scopedRelativePathPrefixes,
-    required this.packageName,
-    required this.allowedPackageImports,
-  });
-
-  final String id;
-  final String message;
-  final DisallowedAstMatchKind kind;
-  final Set<String> cascadedMethodNames;
-  final String? commentSubstring;
-  final Set<String> rawNamedTypeNames;
-  final String? functionName;
-  final int? maxBodyLineSpan;
-  final bool requireWidgetClassExtends;
-  final int? argumentIndex;
-  final Set<String> invocationMethodNames;
-  final Set<String> allowedRelativePaths;
-  final Set<String> scopedRelativePathPrefixes;
-  final String? packageName;
-  final Set<String> allowedPackageImports;
-}
-
-class DisallowedAstViolation {
-  const DisallowedAstViolation({
-    required this.path,
-    required this.line,
-    required this.ruleId,
-    required this.message,
-  });
-
-  final String path;
-  final int line;
-  final String ruleId;
-  final String message;
 }
 
 Expression _unwrapParenthesized(Expression expr) {
